Create the logic for the game
I developed several functions to allow a user to play the game. It begins
by defining an arr of possible choices and then selects one of the choices
from said arr as the computer choice. At the bottom of the code there is
a function that gathers the user's choice. These are the two inputs for
the round function which determines who wins each round. All of these
functions are called during the game function which is currently set to
loop through 5 rounds and log the results of that round to the console.